Fix stability metric

+++ fuzzer/src/shared_state.rs
@@ -22,7 +22,7 @@ pub struct GlobalSharedState {
     //false for not crashing input. True for crashing inputs
     pub bitmaps: HashMap<bool, Vec<u8>>,
     pub execution_count: u64,
-    pub non_determenistic_execution: u64,
+    pub non_determenistic_execution: f64,
     pub average_executions_per_sec: u32,
     pub bits_found_by_havoc: u64,
     pub bits_found_by_havoc_rec: u64,
@@ -58,7 +58,7 @@ impl GlobalSharedState {
             queue,
             bitmaps,
             execution_count: 0,
-            non_determenistic_execution: 0,
+            non_determenistic_execution: 0.0,
             average_executions_per_sec: 0,
             bits_found_by_havoc: 0,
             bits_found_by_havoc_rec: 0,
